optimize creating app and provider, refactor in fn
