+++ central/Cargo.toml
@@ -13,7 +13,6 @@ shared = { workspace = true }
 serde = { workspace = true }
 futures = { workspace = true }
 serde_json = "1"
-rand = "0.9"
 reqwest = { version = "0.12", default_features = false, features = ["default-tls"] }
 anyhow = { workspace = true }
 tracing = { workspace = true }
