+++ shared/Cargo.toml
@@ -8,3 +8,4 @@ license = "Apache-2.0"
 
 [dependencies]
 serde = { workspace = true }
+rand = "0.9.1"
